+++ packages/erxes-ui/src/types.ts
@@ -1,5 +1,5 @@
-import { IActivityLogForMonth } from '@erxes/ui-log/src/activityLogs/types';
-import { IUser } from './auth/types';
+import { IActivityLogForMonth } from "@erxes/ui-log/src/activityLogs/types";
+import { IUser } from "./auth/types";
 
 export interface IRouterProps {
   location: any;
@@ -188,7 +188,7 @@ export type IEditorProps = {
   showMentions?: boolean;
   toolbar?: any[];
   autoFocus?: boolean;
-  toolbarLocation?: 'top' | 'bottom';
+  toolbarLocation?: "top" | "bottom";
   autoGrow?: boolean;
   autoGrowMinHeight?: number | string;
   autoGrowMaxHeight?: number | string;
@@ -196,6 +196,9 @@ export type IEditorProps = {
   isSubmitted?: boolean;
   formItems?: any;
   contentType?: string;
+  additionalToolbarContent?: (props: {
+    onClick: (placeholder: string) => void;
+  }) => React.ReactNode;
 };
 
 export type QueryResponse = {