+++ packages/plugin-ebarimt-ui/src/graphql/mutations.ts
@@ -1,6 +1,6 @@
 // Settings
 
-import { responseFields } from "./queries";
+import { ebarimtProductRuleFields, responseFields } from "./queries";
 
 const updateConfigs = `
   mutation configsUpdate($configsMap: JSON!) {
@@ -24,9 +24,61 @@ const putResponseReReturn = `
   }
 `;
 
+const mutationParamsDef = `
+  $title: String
+  $productIds: [String]
+  $productCategoryIds: [String]
+  $excludeCategoryIds: [String]
+  $excludeProductIds: [String]
+  $tagIds: [String]
+  $excludeTagIds: [String]
+  $kind: String
+  $taxType: String
+  $taxCode: String
+  $taxPercent: Float
+`;
+
+const mutationParamsVal = `
+  title: $title
+  productIds: $productIds
+  productCategoryIds: $productCategoryIds
+  excludeCategoryIds: $excludeCategoryIds
+  excludeProductIds: $excludeProductIds
+  tagIds: $tagIds
+  excludeTagIds: $excludeTagIds
+  kind: $kind
+  taxType: $taxType
+  taxCode: $taxCode
+  taxPercent: $taxPercent
+`;
+
+const ebarimtProductRuleCreate = `
+  mutation ebarimtProductRuleCreate(${mutationParamsDef}) {
+    ebarimtProductRuleCreate(${mutationParamsVal}) {
+      ${ebarimtProductRuleFields}
+    }
+  }
+`;
+
+const ebarimtProductRuleUpdate = `
+  mutation ebarimtProductRuleUpdate($_id: String!, ${mutationParamsDef}) {
+    ebarimtProductRuleUpdate(_id: $_id, ${mutationParamsVal}) {
+      ${ebarimtProductRuleFields}
+    }
+  }
+`;
+
+const ebarimtProductRulesRemove = `
+  mutation ebarimtProductRulesRemove($ids: [String]) {
+    ebarimtProductRulesRemove(ids: $ids)
+  }
+`;
 
 export default {
   updateConfigs,
   putResponseReturnBill,
-  putResponseReReturn
+  putResponseReReturn,
+  ebarimtProductRuleCreate,
+  ebarimtProductRuleUpdate,
+  ebarimtProductRulesRemove,
 };