+++ packages/plugin-knowledgebase-api/src/messageBroker.ts
@@ -57,6 +57,18 @@ export const setupMessageConsumers = async () => {
   );
 
   consumeRPCQueue(
+    "knowledgebase:articles.count",
+    async ({ subdomain, data: { query } }) => {
+      const models = await generateModels(subdomain);
+
+      return {
+        status: "success",
+        data: await models.KnowledgeBaseArticles.countDocuments(query)
+      };
+    }
+  );
+
+  consumeRPCQueue(
     "knowledgebase:categories.findOne",
     async ({ subdomain, data: { query } }) => {
       const models = await generateModels(subdomain);