+++ packages/ui-forms/src/segmenForm.tsx
@@ -22,16 +22,17 @@ export default function Form({ config, onChangeConfig }: Props) {
         queryName="integrations"
         customQuery={queries.integrations}
         filterParams={{ kind: 'lead', perPage: 1000 }}
-        generateOptions={(array) => {
-          if (!config.formId && !array?.length) {
-            onChangeConfig && onChangeConfig({ formId: array[0].form?._id });
+        generateOptions={array => {
+          if (!config.formId && !!array?.length) {
+            const { form = {} } = (array || [])[0] || {};
+            onChangeConfig && onChangeConfig({ formId: form?._id });
           }
-          return array.map((item) => ({
+          return array.map(item => ({
             label: item.name,
             value: item?.form?._id,
           }));
         }}
-        onSelect={(value) => {
+        onSelect={value => {
           onChangeConfig && onChangeConfig({ formId: value });
         }}
       />